+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 
 - Fix accidental double allocation in `ExactLengthArrayPool.OfLength(int)`.
+- Micro optimization by trying to avoid contention when doing `ObjectPool<T>.Shared.Rent()`.
 
 ## v0.3.0
 - Fix documentation typos.